fix: move retrieveTransactionDetails from service to data provider; remove unnecessary getter; move and rework unit tests
